pyproject.toml
[project]
name = "plant-disease-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "numpy>=1.26.0",
    "pillow>=10.0.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.9",
    "tensorflow>=2.15.0",
    "uvicorn[standard]>=0.29.0",
    "python-dotenv>=1.0.0",
    "langchain-core>=0.3.0",
    "langchain-google-genai>=2.1.0",
    "langchain-tavily>=0.1.0",
]
